Feedstocks: When updating feedstocks skip empty repos, but warn user.
In some cases, feedstock generation fails. When this happens, repos
could be created, but have not commits in them. As a result, we can't
add them as a submodule because we don't have a commit to point to. This
resulted in a failure. However, the error message provided was a bit
uninformative for those unfamiliar with the code. Also, it stops adding
other working repos. As there may be cases where we have a broken repo
and we don't want this grinding to a halt, here we raise a warning that
both explains what is going on and let's us proceed through the working
repos.