Suggest conda-forge as fallback channel, not default
`conda config --add channels conda-forge` inserts conda-forge as the first channel to be searched.  This generally leads to the replacement of many default packages the user may not have intended to modify.  (If this is the recommended approach it should be clearly explained)

Using the `--append` argument instead uses conda-forge only for packages not provided in the default channels, which is a more sensible default for new users.  `--append` also clearly adds to the end of a list, where `--add` is ambiguous.