Update feedstocks (forcibly).
It is possible due to formatting issues or other problems that we are
unable to run `git pull` on each submodule. This is because we will need
to overwrite changes in the feedstock to update it. To fix this we need
to fetch the remote and then reset hard to the new location. This makes
that change.