FIX: Update outdated pins
Previously this script was able to update recipes with outdated pins.
However, fixes made to the script to avoid partial dependency name
matches, have had an adverse effect on updating outdated dependency
pinnings. A simple fix would be to just split the pinning from the
dependency's name and compare the name alone as we do here. This should
still avoid partial name matches, but should also update old dependency
pins. There are some subtle pinnings that we don't handle with this
strategy. That being said, we haven't tended to use those particular
forms of pinnings yet. If we do, we might want to take a closer look at
`conda` or `conda-build` to find something more robust for this sort of
parsing and comparison.