Initialize submodules first
Before trying to checkout the `master` branch or updating the submodule
to the latest commit, simply make sure the submodule exists. As they are
not all created when cloning the `feedstocks` repo this is very
important. Once the submodule is populated, we can switch to the branch
setup to track the remote and update the submodule to the latest commit.