Switch bot to @conda-forge-coordinator.
So as not to have the same bot handling AppVeyor builds as the one
handling feedstock updates and team updates, this switches the bot to
one that will be dedicated to handling the latter two. This should limit
the frequency with which AppVeyor builds and statuses are blocked by
GitHub rate limiting.