Drop channel manipulations with conda-smithy
As of `conda-smithy` 2.0.0, no longer do the contents of a user's
`.condarc` impact how `conda-smithy` looks for dependencies. Instead it
looks to `conda-forge.yml` for a section called `channels/sources` (a
nested dictionary), which contains a list of channels to search for
packages from. If that section is not filled out (conda-forge default
use case), then it amounts to having `conda-forge` first and then
`defaults` afterwards. As this is the same configuration used to
download packages in the CIs, it makes sense for this configuration to
be used when determining the packages needed and how to proceed with a
re-rendering.