Drop feedstock updates
Every feedstock is now hooked up to a webservice that listens to every
push. The webservice updates the `feedstocks` repo every time a
feedstock pushes something new to `master`. This allows to better mirror
the various feedstocks in conda-forge and include new ones right after
they are created. Also it allows feedstocks to be updated independently
avoiding many past issues.

With this change, it is no longer necessary to comb through all
feedstocks and push all changes to that repo. Though we may need to
catch the occasional feedstock that we miss, it need not involve this
script and can probably reuse the same webservice and/or a command word
to trigger rechecks on feedstocks. As such, we can drop the feedstocks
update step from the CI on this repo.

By simply focusing on the feedstocks that are changing, we should save a
lot of time both human and CI maintaining it. Further by having updates
for any one feedstock independent from others, we should avoid minor
issues (syntax, formatting, etc.) grinding the whole service to a halt.
Also it should go a long ways to mitigate issues with GitHub rate
limiting. Finally the new service will do a better job of catching new
feedstocks that are added and providing better snapshots of conda-forge
throughout the day. This can be used to make announcement about recent
changes or release.