Use absolute link in feedstocks template for anvil
Fix one more link to be absolute instead of relative in the feedstocks
template as it was missed before. Namely fix the link to the anvil image
to be absolute instead of relative. Should ensure the anvil image gets
loaded properly even if it is being referenced from a location outside
the webpage repo.