Use conda-forge org URL in feedstocks template
Replace all instances of conda-forge's GitHub webpage URL with
conda-forge's org URL. This seems to help with loading resources like
CSS, JavaScript, and images. Not to mention this is a location that we
are redirected to anyways by GitHub. So this should avoid rendering
issues that sometimes occur because of this redirect.