Add some info about how to transition from Anaconda's default channel
A common theme I see when helping users on github is that they have
mixed packages from Anaconda + conda-forge. This really isn't the user's
fault since I still remember the days that we had this as the
recommended workflow.

However, today this workflow leads to totally broken environments.

I would like to document a guide that I can point them to and politely
say:

> Unfortunately this is no longer supported, see .... for how you can
migrate.