Use dagree for interactive migration graph
In addition to the normal graphviz svg
dagree is already a dependency via mermaid/docusaurus

This filter and a remove packages that are "done", and/or have no
parrent children befor showing the graph, and split it into connected
components.

The interactive graph has the following features:
  - span/zoom with mouse/trackpad
  - hover a node to dimm all nodes except direct children/parents.
  - click on a node to show only current graph and all parents (and
    their parents...)/children (and their children...), but not
    siblings.

A filter search bar to find a node easily and select it without having
to click on it.

A dropdown for the various dagree layout options I think are
un-necessary but let us explore the various dagree options.

The node color should reflect the ci status of the PR, but are not
properly clickable to open given PR, this is something I'm planning to
do if there is interest.

Note that there are some hacks as some packages are marked as "awaiting
parent", while they do not have any parent pending (like pyside2 for
314t, llvmlite for 314t also IIRC), maybe this should be fixed in the
table as well with a different label.

I'm also rendering the graph even if there is an extremly large number
of node, I think I there should be a confirmation if there is say > 1k
nodes and a confirmation, but one of my powerful machine handles things
really well so not sure.