Share table sorting component & update sortable keys
- This extract the sorting tables into their own components.
- Use that also in /status/migrations/?name=....

Therefore it adds logic to also sort by CI status and updated_at,
as ci status is an enum, we now track the previous column we sort with
so that you can sort by status (and break tie by updated_at), or any
other tie break with the second to last sorted column.

CI status and updated_at that are null are always sorted last.

I guess the sorting table coudl take a mapping from field name to a
comparison function, but it would be starting to be a bit of a heavy
handed approach I think.